+++ src/test/java/com/jnape/palatable/lambda/semigroup/builtin/AddAllTest.java
@@ -3,21 +3,18 @@ package com.jnape.palatable.lambda.semigroup.builtin;
 import org.junit.Test;
 
 import java.util.HashSet;
-import java.util.Set;
 
-import static com.jnape.palatable.lambda.semigroup.builtin.Concat.concat;
+import static com.jnape.palatable.lambda.semigroup.builtin.AddAll.addAll;
 import static java.util.Collections.singleton;
 import static org.junit.Assert.assertEquals;
 
-public class ConcatTest {
+public class AddAllTest {
 
     @Test
     public void semigroup() {
-        Concat<Integer, Set<Integer>> concat = concat();
-
         assertEquals(new HashSet<Integer>() {{
             add(1);
             add(2);
-        }}, concat.apply(new HashSet<>(singleton(1)), new HashSet<>(singleton(2))));
+        }}, addAll(new HashSet<>(singleton(1)), new HashSet<>(singleton(2))));
     }
 }